+++ cmd/machine-remediation-operator/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
         "//pkg/operator:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
+        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/client/config:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/manager:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/runtime/signals:go_default_library",
@@ -26,14 +27,33 @@ go_binary(
     x_defs = version_x_defs(),
 )
 
+load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
+
+pkg_tar(
+    name = "crds-tar",
+    srcs = ["//manifests:crds"],
+    mode = "0666",
+    package_dir = "data",
+    visibility = ["//visibility:public"],
+)
+
 load(
     "@io_bazel_rules_docker//container:container.bzl",
     "container_image",
 )
 
 container_image(
-    name = "machine-remediation-operator-image",
+    name = "crds-image",
     base = "//:passwd-image",
+    directory = "/",
+    tars = [":crds-tar"],
+    user = "65534",
+    visibility = ["//visibility:public"],
+)
+
+container_image(
+    name = "machine-remediation-operator-image",
+    base = ":crds-image",
     directory = "/usr/bin/",
     entrypoint = ["/usr/bin/machine-remediation-operator"],
     files = [":machine-remediation-operator"],
