+++ manifests/BUILD.bazel
@@ -1,0 +1,7 @@
+filegroup(
+    name = "crds",
+    srcs = glob([
+        "generated/crds/*.yaml",
+    ]),
+    visibility = ["//visibility:public"],
+)
