+++ manifests/release/machine-remediation-operator.yaml.in
@@ -1,5 +1,2 @@
-{{index .GeneratedManifests "machineremediation_v1alpha1_machinedisruptionbudget.yaml"}}
-{{index .GeneratedManifests "machineremediation_v1alpha1_machinehealthcheck.yaml"}}
-{{index .GeneratedManifests "machineremediation_v1alpha1_machineremediation.yaml"}}
 {{index .GeneratedManifests "machineremediation_v1alpha1_machineremediationoperator.yaml"}}
 {{index .GeneratedManifests "machine-remediation-operator.yaml.in"}}
